requirements.txt
Flask~=2.0.2
numpy~=1.22.2
pandas~=1.4.0
Pillow~=9.0.1
Werkzeug~=2.0.3
WTForms~=3.0.1
faiss-cpu==1.7.2